Adjusted lived location for new cookie middleware and related functions
